Add OnlyKey support to bootstrap
Typically I would only do this if I needed to be able to use the OnlyKey
Chrome app: https://github.com/trustcrypto/OnlyKey-Chrome-App.

I did notice though on a fresh install that I wasn't able to use the key
for U2F auth (Google), but after adding this it did work. So adding as a
default step for overall support if user sets `onlykey_support=true`.