Move to new debian/redhat conditonal pattern
- os_family is an ansible fact that works most of time, but with my
  recent change to Pop_OS! that variable doesn't indicate it is a Debian
derivative which means that I'd have to start adding many lines of
compound conditionals to all role task files that do the same switching.
Instead, I now create a group for debian_family and redhat_family that
will include anything of the respective os_family while also including a
list of alternate distros that should be in that family.
- Convert over the primary roles to new pattern and align the task file
  names