Fix weird YAML parsing bug
Noticed with 20.04 install, but for whatever reason was getting issues
with the commented portion of line after the uncommented characters.
Changed the weird single quotes to normal single quotes, but still had
issues, then removed them alltogether and that did it? No idea why but
now no errors when using the yaml inventory plugin.